init inferencer wrapper
