+++ mmedit/apis/diffusers_wrapped/__init__.py
@@ -1,0 +1,5 @@
+# Copyright (c) OpenMMLab. All rights reserved.
+from .diffusers_inferencer import DiffusersInferencer
+from .stable_diffusion import StableDiffusionInferencer
+
+__all__ = ['StableDiffusionInferencer', 'DiffusersInferencer']
