+++ mmedit/apis/diffusers_wrapped/stable_diffusion/__init__.py
@@ -1,0 +1,4 @@
+# Copyright (c) OpenMMLab. All rights reserved.
+from .stable_diffusion_inferencer import StableDiffusionInferencer
+
+__all__ = ['StableDiffusionInferencer']
