+++ mmedit/utils/__init__.py
@@ -1,13 +1,16 @@
 # Copyright (c) OpenMMLab. All rights reserved.
 from .cli import modify_args
 from .img_utils import get_box_info, reorder_image, tensor2img, to_numpy
-from .io_utils import MMEDIT_CACHE_DIR, download_from_url
+from .io_utils import (MMEDIT_CACHE_DIR, download_from_url, read_frames,
+                       read_image)
 # TODO replace with engine's API
 from .logger import print_colored_log
 from .sampler import get_sampler
-from .setup_env import register_all_modules, try_import
+from .setup_env import (delete_cfg, init_model, register_all_modules,
+                        set_random_seed, try_import)
 from .trans_utils import (add_gaussian_noise, adjust_gamma, bbox2mask,
-                          brush_stroke_mask, get_irregular_mask, make_coord,
+                          brush_stroke_mask, calculate_grid_size,
+                          get_irregular_mask, make_coord, pad_sequence,
                           random_bbox, random_choose_unknown)
 from .typing import ConfigType, ForwardInputs, LabelVar, NoiseVar, SampleList
 
@@ -17,5 +20,7 @@ __all__ = [
     'MMEDIT_CACHE_DIR', 'download_from_url', 'get_sampler', 'tensor2img',
     'random_choose_unknown', 'add_gaussian_noise', 'adjust_gamma',
     'make_coord', 'bbox2mask', 'brush_stroke_mask', 'get_irregular_mask',
-    'random_bbox', 'reorder_image', 'to_numpy', 'get_box_info'
+    'random_bbox', 'reorder_image', 'to_numpy', 'get_box_info', 'init_model',
+    'read_image', 'read_frames', 'set_random_seed', 'delete_cfg',
+    'calculate_grid_size', 'pad_sequence'
 ]
