+++ mmedit/utils/io_utils.py
@@ -3,13 +3,18 @@ import hashlib
 import os
 
 import click
+import mmcv
 import mmengine
+import numpy as np
 import requests
 import torch.distributed as dist
 from mmengine.dist import get_dist_info
+from mmengine.fileio import get_file_backend
 from requests.exceptions import InvalidURL, RequestException, Timeout
 
 MMEDIT_CACHE_DIR = os.path.expanduser('~') + '/.cache/openmmlab/mmedit/'
+VIDEO_EXTENSIONS = ('.mp4', '.mov', '.avi')
+FILE_CLIENT = get_file_backend(backend_args={'backend': 'local'})
 
 
 def get_content_from_url(url, timeout=15, stream=False):
@@ -96,3 +101,45 @@ def download_from_url(url,
         dist.barrier()
 
     return dest_path
+
+
+def read_image(filepath):
+    """Read image from file.
+
+    Args:
+        filepath (str): File path.
+
+    Returns:
+        image (np.array): Image.
+    """
+    img_bytes = FILE_CLIENT.get(filepath)
+    image = mmcv.imfrombytes(
+        img_bytes, flag='color', channel_order='rgb', backend='pillow')
+    return image
+
+
+def read_frames(source, start_index, num_frames, from_video, end_index):
+    """Read frames from file or video.
+
+    Args:
+        source (list | mmcv.VideoReader): Source of frames.
+        start_index (int): Start index of frames.
+        num_frames (int): frames number to be read.
+        from_video (bool): Weather read frames from video.
+        end_index (int): The end index of frames.
+
+    Returns:
+        images (np.array): Images.
+    """
+    images = []
+    last_index = min(start_index + num_frames, end_index)
+    # read frames from video
+    if from_video:
+        for index in range(start_index, last_index):
+            if index >= source.frame_cnt:
+                break
+            images.append(np.flip(source.get_frame(index), axis=2))
+    else:
+        files = source[start_index:last_index]
+        images = [read_image(f) for f in files]
+    return images
