+++ mmedit/utils/trans_utils.py
@@ -493,3 +493,55 @@ def random_choose_unknown(unknown, crop_size):
     left = np.clip(center_w - delta_w, 0, w - crop_w)
 
     return top, left
+
+
+def pad_sequence(data, window_size):
+    """Pad frame sequence data.
+
+    Args:
+        data (Tensor): The frame sequence data.
+        window_size (int): The window size used in sliding-window framework.
+
+    Returns:
+        data (Tensor): The padded result.
+    """
+
+    padding = window_size // 2
+
+    data = torch.cat([
+        data[:, 1 + padding:1 + 2 * padding].flip(1), data,
+        data[:, -1 - 2 * padding:-1 - padding].flip(1)
+    ],
+                     dim=1)
+
+    return data
+
+
+def calculate_grid_size(num_batches: int = 1, aspect_ratio: int = 1) -> int:
+    """Calculate the number of images per row (nrow) to make the grid closer to
+    square when formatting a batch of images to grid.
+
+    Args:
+        num_batches (int, optional): Number of images per batch. Defaults to 1.
+        aspect_ratio (int, optional): The aspect ratio (width / height) of
+            each image sample. Defaults to 1.
+
+    Returns:
+        int: Calculated number of images per row.
+    """
+    curr_ncol, curr_nrow = 1, num_batches
+    curr_delta = curr_nrow * aspect_ratio - curr_ncol
+
+    nrow = curr_nrow
+    delta = curr_delta
+
+    while curr_delta > 0:
+
+        curr_ncol += 1
+        curr_nrow = math.ceil(num_batches / curr_ncol)
+
+        curr_delta = curr_nrow * aspect_ratio - curr_ncol
+        if curr_delta < delta and curr_delta >= 0:
+            nrow, delta = curr_nrow, curr_delta
+
+    return nrow
